backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
bidict==0.23.1
black==24.1.1
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
cryptography==45.0.7
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
openai==1.75.0
stripe==12.1.0
email-validator==2.1.0
executing==2.2.1
fastapi==0.109.0
flake8==7.0.0
frozenlist==1.8.0
google-ai-generativelanguage==0.4.0
google-api-core==2.26.0
google-auth==2.41.1
google-generativeai==0.3.2
googleapis-common-protos==1.71.0
gotrue==2.12.4
groq==0.33.0
grpcio==1.75.1
grpcio-status==1.62.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
ipython==8.18.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mccabe==0.7.0
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.5.0
pluggy==1.6.0
postgrest==0.18.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.1.0
pydantic_core==2.41.4
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
PyPDF2==3.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.3
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.11.0
PyYAML==6.0.3
realtime==2.22.0
redis==5.0.1
requests==2.32.5
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.35.1
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
tenacity==9.1.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==14.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.22.0
svix==1.14.0
